# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/api/v1/user/block/{user_id}": {
      "post": {
        "description": "Block a user",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain"
        ],
        "tags": [
          "users"
        ],
        "summary": "Block a user",
        "operationId": "users-block-block-user",
        "parameters": [
          {
            "type": "string",
            "description": "Auth Token",
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "type": "integer",
            "description": "user id you want to block",
            "name": "user_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.UserBlock"
            }
          }
        }
      }
    },
    "/api/v1/user/change-password": {
      "patch": {
        "description": "Change password",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Change password",
        "operationId": "users-change-password",
        "parameters": [
          {
            "type": "string",
            "description": "Auth Token",
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "description": "JSON Request Body",
            "name": "RequestBody",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.ChangePasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          }
        }
      }
    },
    "/api/v1/user/delete/{id}": {
      "delete": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Delete users",
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Delete users",
        "operationId": "users-user-delete",
        "parameters": [
          {
            "type": "string",
            "description": "Auth Token",
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "type": "integer",
            "description": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          }
        }
      }
    },
    "/api/v1/user/edit-user-profile": {
      "patch": {
        "description": "Edit user profile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Edit user profile",
        "operationId": "users-edit-user-profile",
        "parameters": [
          {
            "type": "string",
            "description": "Auth Token",
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "description": "JSON Request Body",
            "name": "RequestBody",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.EditUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          }
        }
      }
    },
    "/api/v1/user/my-profile": {
      "post": {
        "description": "Get my profile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Get my profile",
        "operationId": "users-my-profile",
        "parameters": [
          {
            "type": "string",
            "description": "Auth Token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/helper.HttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/models.User"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/api/v1/user/request-otp": {
      "post": {
        "description": "Request OTP",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Request OTP",
        "operationId": "users-request-otp",
        "parameters": [
          {
            "type": "string",
            "description": "Auth Token",
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "description": "JSON Request Body",
            "name": "RequestBody",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.RequestOTPRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/user/upload-profile-photo": {
      "post": {
        "description": "Upload profile photo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "tags": [
          "users"
        ],
        "summary": "Upload profile photo",
        "operationId": "users-upload-profile-photo",
        "parameters": [
          {
            "type": "string",
            "description": "Auth Token",
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "type": "file",
            "description": "image file",
            "name": "image",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "controllers.ChangePasswordRequest": {
      "type": "object",
      "properties": {
        "otp": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "repeat_password": {
          "type": "string"
        }
      }
    },
    "controllers.EditUser": {
      "type": "object",
      "required": [
        "email",
        "name",
        "phone_number",
        "username"
      ],
      "properties": {
        "address": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "phone_number": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "controllers.EditUserRequest": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/controllers.EditUser"
        }
      }
    },
...
    "controllers.RequestOTPRequest": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string"
        }
      }
    },
...
    "helper.HttpResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "additionalProperties": true
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "integer"
        }
      }
    },
    "models.User": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "bank_accounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.UserBankAccount"
          }
        },
        "company": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "is_partner": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "paket": {
          "type": "integer"
        },
        "password": {
          "type": "string"
        },
        "phone_number": {
          "type": "string"
        },
        "player_id": {
          "type": "string"
        },
        "profile_picture": {
          "type": "string"
        },
        "status_active": {
          "type": "integer"
        },
        "updated_at": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "models.UserBankAccount": {
      "type": "object",
      "properties": {
        "account_name": {
          "type": "string"
        },
        "account_number": {
          "type": "string"
        },
        "bank_id": {
          "type": "integer"
        },
        "bank_name": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "updated_at": {
          "type": "string"
        },
        "user_id": {
          "type": "integer"
        }
      }
    },
    "models.UserBlock": {
      "type": "object",
      "properties": {
        "blocked_user_id": {
          "type": "integer"
        },
        "created_at": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "updated_at": {
          "type": "string"
        },
        "user_id": {
          "type": "integer"
        }
      }
    }
  }
}